pyproject.toml
# Copyright 2025 Altaira Labs
# SPDX-License-Identifier: Apache-2.0

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "omnia-langchain-runtime"
version = "0.1.0"
description = "Omnia LangChain runtime for running LangChain agents with gRPC"
readme = "README.md"
license = "Apache-2.0"
authors = [
    { name = "Altaira Labs", email = "dev@altairalabs.com" }
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    # PromptPack libraries
    "promptpack>=0.1.0",
    "promptpack-langchain>=0.1.0",
    # gRPC
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    # LangChain core and LangGraph
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    # LLM providers
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-ollama>=0.3.0",
    # Utilities
    "pyyaml>=6.0.0",
    "pydantic>=2.0",
    "httpx>=0.27.0",
    # Observability
    "prometheus-client>=0.20.0",
]

[project.optional-dependencies]
redis = [
    "redis>=5.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "grpcio-testing>=1.60.0",
]
all = [
    "omnia-langchain-runtime[redis,dev]",
]

[project.scripts]
omnia-langchain-runtime = "omnia_langchain_runtime.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/omnia_langchain_runtime"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "T20"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
